backend/requirements.txt
fastapi
uvicorn
opencv-python
pillow
numpy
torch
torchvision
transformers
python-multipart
pydantic